Improve several minor QoL issues
1. Templates in the tree now show their index in the popup so you immediately know which ones have a hotkey and which ones don't.
2. Instead of selecting the first leaf, now the first template is selected by default when the GUI is opened.
3. TemplateJTree has been split into multiple files to improve IDE performance (in particular for test classes).
